refactors(booking): centralise booking prisma mapper
